refactor(gpu-brain): remove CPU brain code, GpuBrainCompute — GPU-only brain
